feat: 增加无线传感器通信与故障重试（CPU2 V0.9.0.0 / CPU3 V0.11.0.0 / 无线主机 V4.1）
版本：
- CPU2：V0.9.0.0（历史提交编号）
- CPU3：V0.11.0.0（历史提交编号）
- 无线主机：V4.1（历史工程目录标识）

协议版本/兼容性：
- 涉及 DSM、LTD 或无线传感器通信；协议口径以该历史提交的代码树为准。

本次修改：
- 新增 CPU2 无线传感器通信、超时重试和帧校验。
- 新增无线主机通信模块，并同步 CPU3 状态和参数读取。

验证：
- 提交差异复核：通过；本次仅重写说明，代码树保持不变。

+++ LTD_DISPLAY/Communication/external/DSM_modbus/DSM_SlaveModbus_modbus2.c
@@ -14,7 +14,7 @@
 #include "stateformodbus.h"
 
 extern struct MEASURE_DATA Measure_Data;
-int HoldingRegisterArray1[HOLDREGISTERAMOUNT] = {0}; // 保持寄存器数组
+uint16_t HoldingRegisterArray1[HOLDREGISTERAMOUNT] = {0}; // 保持寄存器数组
 int InputRegisterArray[INPUTREGISTERAMOUNT] = {0};	// 输入寄存器数组
 int MaxNum_Coil;									// 线圈最大有效值
 int MaxNum_HoldingRegister;							// 保持寄存器最大有效值
